# fix: copy closure objects before invocation