every closure call invokes `getUniqueThreadClosure` and operates on that. An abstract class between specific closure classes and the abstract function class is added that declares this method abstractly (fixes #2741).