@JonathanStarup 2021-11-29
The new generated classes:
```
public final class Clo%println%121804 extends Clo1%Obj%Obj {
    public Object clo0;

    public final Cont%Obj invoke() {
        Object var1 = this.clo0;
        Object var2 = this.arg0;
        ((PrintStream)((PrintStream)var1)).println((String)((String)var2));
        this.result = Unit.INSTANCE;
        return null;
    }

    public Clo1%Obj%Obj getUniqueThreadClosure() {
        Clo%println%121804 var10000 = new Clo%println%121804();
        var10000.clo0 = this.clo0;
        return var10000;
    }

    public Clo_println_121804/* $FF was: Clo%println%121804*/() {
    }
}
```
```
public abstract class Clo1%Obj%Obj extends Fn1%Obj%Obj {
    public Clo1_Obj_Obj/* $FF was: Clo1%Obj%Obj*/() {
    }

    public abstract Clo1%Obj%Obj getUniqueThreadClosure();
}
```