@JonathanStarup 2021-11-29
Small benchmarks
benchmark |Master | PR
--- | --- | ---
JvmBackend phase | ? | 1018,8 ms (27,2%)