@magnus-madsen 2021-11-29
Its not so much the compiler throughput, but the actual benchmark running times.