@JonathanStarup 2021-11-29
> Does the benchmark with `parMap` run?

which one?