@magnus-madsen 2021-11-29
> > Does the benchmark with `parMap` run?
> 
> which one?

See https://github.com/flix/flix/issues/2336