@JonathanStarup 2021-11-29
> [This program](https://github.com/flix/flix/issues/2336#issuecomment-980541504) returns `Nil`

Yes, it's supposed to return Nil right?