@JonathanStarup 2021-11-29
Branch | benchmark01 | benchmark02 | benchmark03 | benchmark04 | benchmark05 | benchmark06 | benchmark07 | benchmark08 | benchmark09 | benchmark10 | benchmark11 | benchmark12 | benchmark13 | benchmark14 | benchmark15 | benchmark16 | benchmark17 | benchmark18
:--- | ---: | ---: | -- | ---: | ---: | ---: | ---: | ---:| ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---:
Master | 17.640 ns. | 31.697 ns. | 22.138 ns. | 31.184 ns. | 37.739 ns. | 65.333 ns. | 48.628 ns. | 20.833 ns. | 29.707 ns. | 39.850 ns. | 65.349 ns. | 86.099 ns. | 27.976 ns. | 26.784 ns. | 61.112 ns. | 95.296 ns. | 22.727 ns. | 43.720 ns.
PR | 21.579 ns. | 32.767 ns. | 24.016 ns. | 35.284 ns. | 38.921 ns. | 72.073 ns. | 48.804 ns. | 26.182 ns. | 38.971 ns. | 43.577 ns. | 70.283 ns. | 95.579 ns. | 29.460 ns. | 24.371 ns. | 62.138 ns. | 115.304 ns. | 26.982 ns. | 46.703 ns.